+++ Classes/Frontend/FrontendSimulationInterface.php
@@ -5,13 +5,11 @@ declare(strict_types=1);
 namespace Causal\Oidc\Frontend;
 
 use Psr\Http\Message\ServerRequestInterface;
-use TYPO3\CMS\Frontend\Controller\TypoScriptFrontendController;
+use TYPO3\CMS\Frontend\ContentObject\ContentObjectRenderer;
 
 interface FrontendSimulationInterface
 {
-    public function getTSFE(ServerRequestInterface $originalRequest): TypoScriptFrontendController;
-
-    public function getTypoScriptSetup(ServerRequestInterface $originalRequest, TypoScriptFrontendController $tsfe): array;
+    public function getCObj(ServerRequestInterface $originalRequest): ContentObjectRenderer;
 
     public function cleanupTSFE(): void;
 }