+++ Classes/LoginProvider/OidcLoginProvider.php
@@ -8,29 +8,22 @@ use Psr\Http\Message\ServerRequestInterface;
 use TYPO3\CMS\Backend\Controller\LoginController;
 use TYPO3\CMS\Backend\LoginProvider\LoginProviderInterface;
 use TYPO3\CMS\Core\Page\PageRenderer;
-use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Core\View\ViewInterface;
 use TYPO3\CMS\Fluid\View\FluidViewAdapter;
 
 class OidcLoginProvider implements LoginProviderInterface
 {
-    /**
-     * @var int
-     */
-    public const IDENTIFIER = 1742888452490;
+    public const int IDENTIFIER = 1742888452490;
 
     /**
+     * This method is only here, because the Interface in v13
+     * requires it. But the Core actually never uses this method
+     * as long as modifyView is present.
+     *
      * @inheritDoc
      * @param \TYPO3\CMS\Fluid\View\StandaloneView $view
      */
-    public function render($view, PageRenderer $pageRenderer, LoginController $loginController)
-    {
-        $view->setTemplatePathAndFilename(
-            GeneralUtility::getFileAbsFileName('EXT:oidc/Resources/Private/Templates/Backend/LoginOidc.html')
-        );
-
-        $view->assign('enablePasswordReset', false);
-    }
+    public function render($view, PageRenderer $pageRenderer, LoginController $loginController) {}
 
     public function modifyView(ServerRequestInterface $request, ViewInterface $view): string
     {