+++ Classes/Middleware/OauthCallback.php
@@ -51,14 +51,14 @@ class OauthCallback implements MiddlewareInterface, LoggerAwareInterface
 
         if (!$state) {
             $this->logger->debug('Missing state');
-            return (new Response())->withStatus(400, 'Missing state');
+            return new Response()->withStatus(400, 'Missing state');
         }
 
         // A code was supplied, we start the OIDC handling
         $authContext = $this->authenticationContextService->resolveAuthenticationContext($request);
         if (!$authContext) {
             $this->logger->debug('Missing OIDC authentication context', ['request' => $request]);
-            return (new Response())->withStatus(400, 'Missing OIDC authentication context');
+            return new Response()->withStatus(400, 'Missing OIDC authentication context');
         }
         $this->logger->debug('Authentication context is available', ['data' => $authContext]);
 
@@ -74,7 +74,7 @@ class OauthCallback implements MiddlewareInterface, LoggerAwareInterface
                         'error_description' => $queryParams['error_description'] ?? '',
                         'error_uri' => $queryParams['error_uri'] ?? '',
                     ]);
-                    return (new Response())->withStatus(400, 'Communication error with IdP');
+                    return new Response()->withStatus(400, 'Communication error with IdP');
             }
         }
 
@@ -86,7 +86,7 @@ class OauthCallback implements MiddlewareInterface, LoggerAwareInterface
                     'expected' => $authContext->state,
                     'actual' => $state,
                 ]);
-                return (new Response())->withStatus(400, 'Invalid state');
+                return new Response()->withStatus(400, 'Invalid state');
             }
             $this->logger->info('State mismatch. Bypassing CSRF attack mitigation protection according to the extension configuration', [
                 'expected' => $authContext->state,