+++ scripts/phpcs.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-PHP_VERSION=8.2
+PHP_VERSION=8.4
 TOOL_DIR=.Build/tools/phpcs
 TOOL_PACKAGE="friendsofphp/php-cs-fixer"
 TOOL_COMMAND="php-cs-fixer fix -v --diff"